consolidate keyBy ops into associateTasksWithAppeals; update test
