+++ client/app/queue/utils.js
@@ -14,9 +14,15 @@ export const associateTasksWithAppeals = (serverData = {}) => {
       head();
   });
 
+  const tasksById = _.keyBy(tasks, 'id');
+  const appealsById = _(appeals).
+    map((appeal) => ({ ...appeal, docCount: 0 })).
+    keyBy('id').
+    value();
+
   return {
-    appeals,
-    tasks
+    appeals: appealsById,
+    tasks: tasksById
   };
 };
