+++ client/test/node/queue/QueueLoadingScreen-test.js
@@ -1,5 +1,4 @@
 import { expect } from 'chai';
-import _ from 'lodash';
 import { associateTasksWithAppeals, sortTasks } from '../../../app/queue/utils';
 
 const serverData = {
@@ -54,45 +53,44 @@ describe('QueueLoadingScreen', () => {
   it('associates queue decisions/appeals and tasks', () => {
     const { tasks: tasksWithAppeals } = associateTasksWithAppeals(serverData);
 
-    expect(tasksWithAppeals).to.deep.equal([{
-      id: '111',
-      appealId: '123',
-      attributes: {
-        appeal_id: '1',
-        docket_date: '2017-12-28T17:18:20.412Z'
+    expect(tasksWithAppeals).to.deep.equal({
+      111: {
+        id: '111',
+        appealId: '123',
+        attributes: {
+          appeal_id: '1',
+          docket_date: '2017-12-28T17:18:20.412Z'
+        }
+      },
+      222: {
+        id: '222',
+        appealId: '123',
+        attributes: {
+          appeal_id: '1',
+          docket_date: '2016-10-07T03:15:27.580Z'
+        }
+      },
+      333: {
+        id: '333',
+        appealId: '234',
+        attributes: {
+          appeal_id: '2',
+          docket_date: '2015-10-13T06:47:34.155Z'
+        }
+      },
+      444: {
+        id: '444',
+        appealId: '345',
+        attributes: {
+          appeal_id: '3',
+          docket_date: '2016-03-01T04:15:51.123Z'
+        }
       }
-    }, {
-      id: '222',
-      appealId: '123',
-      attributes: {
-        appeal_id: '1',
-        docket_date: '2016-10-07T03:15:27.580Z'
-      }
-    }, {
-      id: '333',
-      appealId: '234',
-      attributes: {
-        appeal_id: '2',
-        docket_date: '2015-10-13T06:47:34.155Z'
-      }
-    }, {
-      id: '444',
-      appealId: '345',
-      attributes: {
-        appeal_id: '3',
-        docket_date: '2016-03-01T04:15:51.123Z'
-      }
-    }]);
+    });
   });
 
   it('groups tasks by AOD/CAVC and sorts by docket date', () => {
-    let { tasks, appeals } = associateTasksWithAppeals(serverData);
-
-    tasks = _.keyBy(tasks, 'id');
-    appeals = _(appeals).
-      map((appeal) => _.extend(appeal, { docCount: 0 })).
-      keyBy('id').
-      value();
+    const { tasks, appeals } = associateTasksWithAppeals(serverData);
 
     const sortedTasks = sortTasks({
       tasks,
